TPM feature
This splits out the TPM functionality into a
feature and permits enrollment/PRT requests using
a simple key pair.
