+++ src/auth.rs
@@ -2,35 +2,91 @@ use crate::error::{ErrorResponse, MsalError};
 use base64::engine::general_purpose::URL_SAFE_NO_PAD;
 use base64::Engine;
 use reqwest::{header, Client};
-use serde::{Deserialize, Deserializer};
+use serde::{Serialize, Deserialize, Deserializer};
 use serde_json::{from_str as json_from_str, Value};
 use urlencoding::encode as url_encode;
 use uuid::Uuid;
 
-#[cfg(feature = "broker")]
-#[doc(cfg(feature = "broker"))]
-use crate::enroll::register_device;
-#[cfg(feature = "broker")]
-#[doc(cfg(feature = "broker"))]
+#[cfg(all(feature = "broker", not(feature = "tpm")))]
+#[doc(cfg(all(feature = "broker", not(feature = "tpm"))))]
+use compact_jwt::crypto::JwsRs256Signer;
+#[cfg(all(feature = "broker", feature = "tpm"))]
+#[doc(cfg(all(feature = "broker", feature = "tpm")))]
 use compact_jwt::crypto::JwsTpmSigner;
 #[cfg(feature = "broker")]
 #[doc(cfg(feature = "broker"))]
 use compact_jwt::jws::JwsBuilder;
-#[cfg(feature = "broker")]
-#[doc(cfg(feature = "broker"))]
+#[cfg(all(feature = "broker", feature = "tpm"))]
+#[doc(cfg(all(feature = "broker", feature = "tpm")))]
 use compact_jwt::traits::JwsMutSigner;
+#[cfg(all(feature = "broker", not(feature = "tpm")))]
+#[doc(cfg(all(feature = "broker", not(feature = "tpm"))))]
+use compact_jwt::JwsSigner;
 #[cfg(feature = "broker")]
 #[doc(cfg(feature = "broker"))]
 use compact_jwt::Jws;
+#[cfg(all(feature = "broker", feature = "tpm"))]
+#[doc(cfg(all(feature = "broker", feature = "tpm")))]
+use kanidm_hsm_crypto::{BoxedDynTpm, IdentityKey, LoadableIdentityKey, MachineKey, Tpm};
+#[cfg(all(feature = "broker", not(feature = "tpm")))]
+#[doc(cfg(all(feature = "broker", not(feature = "tpm"))))]
+use openssl::hash::MessageDigest;
 #[cfg(feature = "broker")]
 #[doc(cfg(feature = "broker"))]
-use kanidm_hsm_crypto::{BoxedDynTpm, IdentityKey, LoadableIdentityKey, MachineKey};
+use openssl::pkey::Public;
+#[cfg(all(feature = "broker", not(feature = "tpm")))]
+#[doc(cfg(all(feature = "broker", not(feature = "tpm"))))]
+use openssl::pkey::{PKey, Private};
+#[cfg(feature = "broker")]
+#[doc(cfg(feature = "broker"))]
+use openssl::rsa::Rsa;
+#[cfg(feature = "broker")]
+#[doc(cfg(feature = "broker"))]
+use openssl::x509::X509;
+#[cfg(all(feature = "broker", not(feature = "tpm")))]
+#[doc(cfg(all(feature = "broker", not(feature = "tpm"))))]
+use openssl::x509::{X509NameBuilder, X509ReqBuilder};
 #[cfg(feature = "broker")]
 #[doc(cfg(feature = "broker"))]
 use os_release::OsRelease;
 #[cfg(feature = "broker")]
 #[doc(cfg(feature = "broker"))]
-use serde::Serialize;
+use reqwest::Url;
+#[cfg(feature = "broker")]
+#[doc(cfg(feature = "broker"))]
+use tracing::debug;
+
+#[cfg(feature = "broker")]
+#[doc(cfg(feature = "broker"))]
+use crate::discovery::{
+    discover_enrollment_services, DISCOVERY_URL, DRS_CLIENT_NAME_HEADER_FIELD,
+    DRS_CLIENT_VERSION_HEADER_FIELD,
+};
+#[cfg(feature = "broker")]
+#[doc(cfg(feature = "broker"))]
+use base64::engine::general_purpose::{STANDARD, URL_SAFE};
+#[cfg(feature = "broker")]
+#[doc(cfg(feature = "broker"))]
+use serde_json::{json, to_string_pretty};
+
+#[cfg(feature = "broker")]
+#[doc(cfg(feature = "broker"))]
+#[derive(Debug, Deserialize)]
+struct Certificate {
+    #[serde(rename = "RawBody")]
+    raw_body: String,
+}
+
+#[cfg(feature = "broker")]
+#[doc(cfg(feature = "broker"))]
+#[derive(Debug, Deserialize)]
+struct DRSResponse {
+    #[serde(rename = "Certificate")]
+    certificate: Certificate,
+}
+
+#[derive(Serialize, Clone, Default)]
+struct JoinPayload {}
 
 #[cfg(feature = "broker")]
 #[doc(cfg(feature = "broker"))]
@@ -565,14 +621,16 @@ impl BrokerClientApplication {
     ///
     /// * `tpm` - The tpm object.
     ///
-    /// * `id_key` - A LoadableIdentityKey which will be used to create the CSR
-    ///   for enrolling the device.
+    /// * `loadable_id_key` - A LoadableIdentityKey which will be used to
+    ///   create the CSR and transport key for enrolling the device.
     ///
     /// # Returns
     ///
     /// * Success: The `id_key` (which has been loaded with a signed
     ///   certificate), and a `device_id`.
     /// * Failure: An MsalError, indicating the failure.
+    #[cfg(feature = "tpm")]
+    #[doc(cfg(feature = "tpm"))]
     pub async fn enroll_device(
         &self,
         username: &str,
@@ -580,12 +638,258 @@ impl BrokerClientApplication {
         domain: &str,
         machine_key: &MachineKey,
         tpm: &mut BoxedDynTpm,
-        id_key: &LoadableIdentityKey,
+        loadable_id_key: &LoadableIdentityKey,
     ) -> Result<(LoadableIdentityKey, String), MsalError> {
         let token = self
             .acquire_token_for_device_enrollment(username, password)
             .await?;
-        register_device(&token.access_token, domain, machine_key, tpm, id_key).await
+
+        // Create the CSR
+        let csr_der = match tpm.identity_key_certificate_request(
+            machine_key,
+            loadable_id_key,
+            "7E980AD9-B86D-4306-9425-9AC066FB014A",
+        ) {
+            Ok(csr_der) => csr_der,
+            Err(e) => return Err(MsalError::TPMFail(format!("Failed creating CSR: {:?}", e))),
+        };
+
+        // Load the transport key
+        let id_key = match tpm.identity_key_load(machine_key, loadable_id_key) {
+            Ok(id_key) => id_key,
+            Err(e) => {
+                return Err(MsalError::TPMFail(format!(
+                    "Failed loading id key: {:?}",
+                    e
+                )))
+            }
+        };
+        let transport_key_der = match tpm.identity_key_public_as_der(&id_key) {
+            Ok(transport_key_pem) => transport_key_pem,
+            Err(e) => {
+                return Err(MsalError::TPMFail(format!(
+                    "Failed getting transport key as der: {:?}",
+                    e
+                )))
+            }
+        };
+        let transport_key_rsa = Rsa::public_key_from_der(&transport_key_der)
+            .map_err(|e| MsalError::TPMFail(format!("{}", e)))?;
+
+        let (cert, device_id) = self
+            .enroll_device_internal(&token.access_token, domain, &transport_key_rsa, &csr_der)
+            .await?;
+
+        let new_loadable_id_key = match tpm.identity_key_associate_certificate(
+            machine_key,
+            loadable_id_key,
+            &cert
+                .to_der()
+                .map_err(|e| MsalError::TPMFail(format!("{}", e)))?,
+        ) {
+            Ok(loadable_id_key) => loadable_id_key,
+            Err(e) => {
+                return Err(MsalError::TPMFail(format!(
+                    "Failed creating loadable identity key: {:?}",
+                    e
+                )))
+            }
+        };
+
+        Ok((new_loadable_id_key, device_id.to_string()))
+    }
+
+    /// Enroll the device in the directory.
+    ///
+    /// # Arguments
+    ///
+    /// * `username` - Typically a UPN in the form of an email address.
+    ///
+    /// * `password` - The password.
+    ///
+    /// * `domain` - The domain the device is to be enrolled in.
+    ///
+    /// * `transport_key` - An RSA private key which will be used to
+    ///   create the CSR and transport key for enrolling the device.
+    ///
+    /// # Returns
+    ///
+    /// * Success: An x509 certificate signed during enrollment, and a
+    ///   `device_id`.
+    /// * Failure: An MsalError, indicating the failure.
+    #[cfg(not(feature = "tpm"))]
+    #[doc(cfg(not(feature = "tpm")))]
+    pub async fn enroll_device(
+        &self,
+        username: &str,
+        password: &str,
+        domain: &str,
+        id_key: &PKey<Private>,
+    ) -> Result<(X509, String), MsalError> {
+        let token = self
+            .acquire_token_for_device_enrollment(username, password)
+            .await?;
+
+        let mut req_builder =
+            X509ReqBuilder::new().map_err(|e| MsalError::TPMFail(format!("{}", e)))?;
+
+        let mut x509_name =
+            X509NameBuilder::new().map_err(|e| MsalError::TPMFail(format!("{}", e)))?;
+
+        x509_name
+            .append_entry_by_text("CN", "7E980AD9-B86D-4306-9425-9AC066FB014A")
+            .map_err(|e| MsalError::TPMFail(format!("{}", e)))?;
+
+        let x509_name = x509_name.build();
+        req_builder
+            .set_subject_name(&x509_name)
+            .map_err(|e| MsalError::TPMFail(format!("{}", e)))?;
+
+        req_builder
+            .set_pubkey(id_key)
+            .map_err(|e| MsalError::TPMFail(format!("{}", e)))?;
+
+        req_builder
+            .sign(id_key, MessageDigest::sha256())
+            .map_err(|e| MsalError::TPMFail(format!("{}", e)))?;
+
+        let csr_der = req_builder
+            .build()
+            .to_der()
+            .map_err(|e| MsalError::TPMFail(format!("{}", e)))?;
+
+        let public_key = Rsa::public_key_from_der(
+            &id_key
+                .raw_public_key()
+                .map_err(|e| MsalError::DeviceEnrollmentFail(format!("{}", e)))?,
+        )
+        .map_err(|e| MsalError::DeviceEnrollmentFail(format!("{}", e)))?;
+
+        self.enroll_device_internal(&token.access_token, domain, &public_key, &csr_der)
+            .await
+    }
+
+    async fn enroll_device_internal(
+        &self,
+        access_token: &str,
+        domain: &str,
+        transport_key: &Rsa<Public>,
+        csr_der: &Vec<u8>,
+    ) -> Result<(X509, String), MsalError> {
+        let enrollment_services =
+            discover_enrollment_services(&self.app.client(), access_token, domain).await?;
+        let (join_endpoint, service_version) = match enrollment_services.device_join_service {
+            Some(device_join_service) => {
+                let join_endpoint = match device_join_service.endpoint {
+                    Some(join_endpoint) => join_endpoint,
+                    None => format!("{}/EnrollmentServer/device/", DISCOVERY_URL).to_string(),
+                };
+                let service_version = match device_join_service.service_version {
+                    Some(service_version) => service_version,
+                    None => "2.0".to_string(),
+                };
+                (join_endpoint, service_version)
+            }
+            None => (
+                format!("{}/EnrollmentServer/device/", DISCOVERY_URL).to_string(),
+                "2.0".to_string(),
+            ),
+        };
+
+        let url = Url::parse_with_params(&join_endpoint, &[("api-version", service_version)])
+            .map_err(|e| MsalError::URLFormatFailed(format!("{}", e)))?;
+
+        let host: String = match hostname::get()
+            .map_err(|e| MsalError::GeneralFailure(format!("{}", e)))?
+            .to_str()
+        {
+            Some(host) => String::from(host),
+            None => {
+                return Err(MsalError::GeneralFailure(
+                    "Failed to get machine hostname for enrollment".to_string(),
+                ))
+            }
+        };
+
+        let os_release =
+            OsRelease::new().map_err(|e| MsalError::GeneralFailure(format!("{}", e)))?;
+
+        let transport_key_der = transport_key
+            .public_key_to_der()
+            .map_err(|e| MsalError::DeviceEnrollmentFail(format!("{}", e)))?;
+        let jwk = json!({
+            "kty": "RSA",
+            "kid": Uuid::new_v4(),
+            "e": URL_SAFE_NO_PAD.encode(transport_key.e().to_vec()),
+            "n": URL_SAFE_NO_PAD.encode(transport_key_der)
+        });
+        let encoded_stk = URL_SAFE.encode(jwk.to_string());
+
+        let payload = json!({
+            "CertificateRequest": {
+                "Type": "pkcs10",
+                "Data": STANDARD.encode(csr_der)
+            },
+            "DeviceDisplayName": host,
+            "DeviceType": "Linux",
+            "JoinType": 0,
+            "OSVersion": format!("{} {}", os_release.pretty_name, os_release.version_id),
+            "TargetDomain": domain,
+            "TransportKey": encoded_stk,
+            "Attributes": {
+                "ReuseDevice": "true",
+                "ReturnClientSid": "true"
+            }
+        });
+        if let Ok(pretty) = to_string_pretty(&payload) {
+            debug!("POST {}: {}", url, pretty);
+        }
+        let resp = self
+            .app
+            .client()
+            .post(url)
+            .header(header::AUTHORIZATION, format!("Bearer {}", access_token))
+            .header(header::CONTENT_TYPE, "application/json")
+            .header(DRS_CLIENT_NAME_HEADER_FIELD, env!("CARGO_PKG_NAME"))
+            .header(DRS_CLIENT_VERSION_HEADER_FIELD, env!("CARGO_PKG_VERSION"))
+            .header(header::ACCEPT, "application/json, text/plain, */*")
+            .json(&payload)
+            .send()
+            .await
+            .map_err(|e| MsalError::RequestFailed(format!("{}", e)))?;
+        if resp.status().is_success() {
+            let res: DRSResponse = resp
+                .json()
+                .await
+                .map_err(|e| MsalError::InvalidJson(format!("{}", e)))?;
+            let cert = X509::from_pem(
+                format!(
+                    "-----BEGIN CERTIFICATE-----\n{}\n-----END CERTIFICATE-----",
+                    res.certificate.raw_body
+                )
+                .as_bytes(),
+            )
+            .map_err(|e| MsalError::GeneralFailure(format!("{}", e)))?;
+            let subject_name = cert.subject_name();
+            let device_id = match subject_name.entries().next() {
+                Some(entry) => entry
+                    .data()
+                    .as_utf8()
+                    .map_err(|e| MsalError::GeneralFailure(format!("{}", e)))?,
+                None => {
+                    return Err(MsalError::GeneralFailure(
+                        "The device id was missing from the certificate response".to_string(),
+                    ))
+                }
+            };
+            Ok((cert, device_id.to_string()))
+        } else {
+            Err(MsalError::GeneralFailure(
+                resp.text()
+                    .await
+                    .map_err(|e| MsalError::GeneralFailure(format!("{}", e)))?,
+            ))
+        }
     }
 
     /// Gets a token for a given resource via user credentials.
@@ -706,6 +1010,25 @@ impl BrokerClientApplication {
         }
     }
 
+    pub async fn build_jwt_by_username_password(
+        &self,
+        username: &str,
+        password: &str,
+    ) -> Result<Jws, MsalError> {
+        let nonce = self.request_nonce().await?;
+
+        Ok(JwsBuilder::from(
+            serde_json::to_vec(&UsernamePasswordAuthenticationPayload::new(
+                username, password, &nonce,
+            ))
+            .map_err(|e| {
+                MsalError::InvalidJson(format!("Failed serializing UsernamePassword JWT: {}", e))
+            })?,
+        )
+        .set_typ(Some("JWT"))
+        .build())
+    }
+
     /// Gets a Primary Refresh Token (PRT) via user credentials.
     ///
     /// # Arguments
@@ -721,6 +1044,8 @@ impl BrokerClientApplication {
     /// # Returns
     /// * Success: A PrimaryRefreshToken, containing a refresh_token and tgt.
     /// * Failure: An MsalError, indicating the failure.
+    #[cfg(feature = "tpm")]
+    #[doc(cfg(feature = "tpm"))]
     pub async fn acquire_user_prt_by_username_password(
         &self,
         username: &str,
@@ -728,20 +1053,57 @@ impl BrokerClientApplication {
         tpm: &mut BoxedDynTpm,
         id_key: &IdentityKey,
     ) -> Result<PrimaryRefreshToken, MsalError> {
+        let jwt = self
+            .build_jwt_by_username_password(username, password)
+            .await?;
+        let signed_jwt = self.sign_jwt(&jwt, tpm, id_key).await?;
+
+        self.acquire_user_prt_jwt(&signed_jwt).await
+    }
+
+    /// Gets a Primary Refresh Token (PRT) via user credentials.
+    ///
+    /// # Arguments
+    ///
+    /// * `username` - Typically a UPN in the form of an email address.
+    ///
+    /// * `password` - The password.
+    ///
+    /// * `id_key` - The private key used during device enrollment.
+    ///
+    /// # Returns
+    /// * Success: A PrimaryRefreshToken, containing a refresh_token and tgt.
+    /// * Failure: An MsalError, indicating the failure.
+    #[cfg(not(feature = "tpm"))]
+    #[doc(cfg(not(feature = "tpm")))]
+    pub async fn acquire_user_prt_by_username_password(
+        &self,
+        username: &str,
+        password: &str,
+        id_key: &PKey<Private>,
+    ) -> Result<PrimaryRefreshToken, MsalError> {
+        let jwt = self
+            .build_jwt_by_username_password(username, password)
+            .await?;
+        let signed_jwt = self.sign_jwt(&jwt, id_key).await?;
+
+        self.acquire_user_prt_jwt(&signed_jwt).await
+    }
+
+    async fn build_jwt_by_refresh_token(&self, refresh_token: &str) -> Result<Jws, MsalError> {
         let nonce = self.request_nonce().await?;
 
-        let jwt = JwsBuilder::from(
-            serde_json::to_vec(&UsernamePasswordAuthenticationPayload::new(
-                username, password, &nonce,
+        Ok(JwsBuilder::from(
+            serde_json::to_vec(&RefreshTokenAuthenticationPayload::new(
+                refresh_token,
+                &nonce,
             ))
             .map_err(|e| {
-                MsalError::InvalidJson(format!("Failed serializing UsernamePassword JWT: {}", e))
+                MsalError::InvalidJson(format!("Failed serializing RefreshToken JWT: {}", e))
             })?,
         )
         .set_typ(Some("JWT"))
-        .build();
-
-        self.acquire_user_prt_jwt(&jwt, tpm, id_key).await
+        .build())
     }
 
     /// Gets a Primary Refresh Token (PRT) via a refresh token (RT) obtained
@@ -760,36 +1122,55 @@ impl BrokerClientApplication {
     /// # Returns
     /// * Success: A PrimaryRefreshToken, containing a refresh_token and tgt.
     /// * Failure: An MsalError, indicating the failure.
+    #[cfg(feature = "tpm")]
+    #[doc(cfg(feature = "tpm"))]
     pub async fn acquire_user_prt_by_refresh_token(
         &self,
         refresh_token: &str,
         tpm: &mut BoxedDynTpm,
         id_key: &IdentityKey,
     ) -> Result<PrimaryRefreshToken, MsalError> {
-        let nonce = self.request_nonce().await?;
+        let jwt = self.build_jwt_by_refresh_token(refresh_token).await?;
+        let signed_jwt = self.sign_jwt(&jwt, tpm, id_key).await?;
 
-        let jwt = JwsBuilder::from(
-            serde_json::to_vec(&RefreshTokenAuthenticationPayload::new(
-                refresh_token,
-                &nonce,
-            ))
-            .map_err(|e| {
-                MsalError::InvalidJson(format!("Failed serializing RefreshToken JWT: {}", e))
-            })?,
-        )
-        .set_typ(Some("JWT"))
-        .build();
-
-        self.acquire_user_prt_jwt(&jwt, tpm, id_key).await
+        self.acquire_user_prt_jwt(&signed_jwt).await
     }
 
-    async fn acquire_user_prt_jwt(
+    /// Gets a Primary Refresh Token (PRT) via a refresh token (RT) obtained
+    /// previously.
+    ///
+    /// # Arguments
+    ///
+    /// * `username` - Typically a UPN in the form of an email address.
+    ///
+    /// * `password` - The password.
+    ///
+    /// * `id_key` - The private key used during device enrollment.
+    ///
+    /// # Returns
+    /// * Success: A PrimaryRefreshToken, containing a refresh_token and tgt.
+    /// * Failure: An MsalError, indicating the failure.
+    #[cfg(not(feature = "tpm"))]
+    #[doc(cfg(not(feature = "tpm")))]
+    pub async fn acquire_user_prt_by_refresh_token(
+        &self,
+        refresh_token: &str,
+        id_key: &PKey<Private>,
+    ) -> Result<PrimaryRefreshToken, MsalError> {
+        let jwt = self.build_jwt_by_refresh_token(refresh_token).await?;
+        let signed_jwt = self.sign_jwt(&jwt, id_key).await?;
+
+        self.acquire_user_prt_jwt(&signed_jwt).await
+    }
+
+    #[cfg(feature = "tpm")]
+    #[doc(cfg(feature = "tpm"))]
+    async fn sign_jwt(
         &self,
         jwt: &Jws,
         tpm: &mut BoxedDynTpm,
         id_key: &IdentityKey,
-    ) -> Result<PrimaryRefreshToken, MsalError> {
-        // [MS-OAPXBC] 3.2.5.1.2 POST (Request for Primary Refresh Token)
+    ) -> Result<String, MsalError> {
         let mut jws_tpm_signer = match JwsTpmSigner::new(tpm, id_key) {
             Ok(jws_tpm_signer) => jws_tpm_signer,
             Err(e) => {
@@ -805,10 +1186,43 @@ impl BrokerClientApplication {
             Err(e) => return Err(MsalError::TPMFail(format!("Failed signing jwk: {}", e))),
         };
 
+        Ok(format!("{}", signed_jwt))
+    }
+
+    #[cfg(not(feature = "tpm"))]
+    #[doc(cfg(not(feature = "tpm")))]
+    async fn sign_jwt(&self, jwt: &Jws, id_key: &PKey<Private>) -> Result<String, MsalError> {
+        let jws_rs256_signer = match JwsRs256Signer::from_rs256_der(
+            &id_key
+                .private_key_to_der()
+                .map_err(|e| MsalError::CryptoFail(format!("{}", e)))?,
+        ) {
+            Ok(jws_rs256_signer) => jws_rs256_signer,
+            Err(e) => {
+                return Err(MsalError::CryptoFail(format!(
+                    "Failed loading rs256 signer: {}",
+                    e
+                )))
+            }
+        };
+
+        let signed_jwt = match jws_rs256_signer.sign(jwt) {
+            Ok(signed_jwt) => signed_jwt,
+            Err(e) => return Err(MsalError::CryptoFail(format!("Failed signing jwk: {}", e))),
+        };
+
+        Ok(format!("{}", signed_jwt))
+    }
+
+    async fn acquire_user_prt_jwt(
+        &self,
+        signed_jwt: &str,
+    ) -> Result<PrimaryRefreshToken, MsalError> {
+        // [MS-OAPXBC] 3.2.5.1.2 POST (Request for Primary Refresh Token)
         let params = [
             ("windows_api_version", "2.0"),
             ("grant_type", "urn:ietf:params:oauth:grant-type:jwt-bearer"),
-            ("request", &format!("{}", signed_jwt)),
+            ("request", &signed_jwt),
             ("client_info", "1"),
             ("tgt", "true"),
         ];
