+++ src/lib.rs
@@ -19,8 +19,4 @@ pub use auth::*;
 
 #[cfg(feature = "broker")]
 #[doc(cfg(feature = "broker"))]
-pub(crate) mod enroll;
-
-#[cfg(feature = "broker")]
-#[doc(cfg(feature = "broker"))]
 pub(crate) mod discovery;
